+++ DATA/BRONZE/StockData/IRCTC_2024_08.csv
@@ -15,4 +15,4 @@ date,open,high,low,close,volume,dividends,stock_splits
 2024-08-21 00:00:00+05:30,922.6,934.75,922.6,930.62,125995,0.0,0.0
 2024-08-22 00:00:00+05:30,933.96,938.24,930.02,935.15,35645,0.0,0.0
 2024-08-23 00:00:00+05:30,935.0,939.1,921.8,923.4,53542,4.0,0.0
-2024-08-26 00:00:00+05:30,924.75,930.95,924.75,927.8,79657,0.0,0.0
+2024-08-26 00:00:00+05:30,924.75,930.95,924.75,928.5,99231,0.0,0.0
