+++ DATA/BRONZE/StockData/KPITTECH_2024_08.csv
@@ -15,4 +15,4 @@ date,open,high,low,close,volume,dividends,stock_splits
 2024-08-21 00:00:00+05:30,1849.95,1855.0,1818.6,1830.15,50454,0.0,0.0
 2024-08-22 00:00:00+05:30,1845.05,1868.1,1827.35,1830.7,20829,0.0,0.0
 2024-08-23 00:00:00+05:30,1854.95,1854.95,1808.2,1824.75,20172,0.0,0.0
-2024-08-26 00:00:00+05:30,1846.95,1861.0,1825.3,1827.85,10494,0.0,0.0
+2024-08-26 00:00:00+05:30,1846.95,1861.0,1825.3,1828.8,12216,0.0,0.0
