+++ DATA/BRONZE/StockData/TATAMOTORS_2024_08.csv
@@ -15,4 +15,4 @@ date,open,high,low,close,volume,dividends,stock_splits
 2024-08-21 00:00:00+05:30,1091.4,1095.1,1080.0,1085.1,228030,0.0,0.0
 2024-08-22 00:00:00+05:30,1090.7,1094.0,1065.05,1068.6,236226,0.0,0.0
 2024-08-23 00:00:00+05:30,1071.8,1092.3,1071.0,1084.9,774166,0.0,0.0
-2024-08-26 00:00:00+05:30,1090.0,1103.0,1083.0,1096.9,232571,0.0,0.0
+2024-08-26 00:00:00+05:30,1090.0,1103.0,1083.0,1093.6,282346,0.0,0.0
