+++ DATA/BRONZE/StockData/TATAPOWER_2024_08.csv
@@ -15,4 +15,4 @@ date,open,high,low,close,volume,dividends,stock_splits
 2024-08-21 00:00:00+05:30,427.0,427.0,421.75,424.3,441380,0.0,0.0
 2024-08-22 00:00:00+05:30,426.0,426.05,421.9,422.85,199822,0.0,0.0
 2024-08-23 00:00:00+05:30,423.0,423.0,417.4,418.0,228889,0.0,0.0
-2024-08-26 00:00:00+05:30,420.1,421.1,417.0,418.0,420545,0.0,0.0
+2024-08-26 00:00:00+05:30,420.1,421.1,417.0,418.85,522246,0.0,0.0
