+++ DATA/GOLD/Holdings/HoldingsTrands_data.csv
@@ -751,10 +751,10 @@ date,open,high,low,close,holding
 2022-05-10,77351.5,78944.63,74896.86,75400.74,59061.4
 2022-05-11,105383.04,107323.17,102456.04,105467.31,88845.4
 2022-05-12,104710.79,104968.16,100188.23,101405.86,88845.4
-2022-05-13,107705.21,111726.46,106256.2,108444.03,102405.4
-2022-05-14,107705.21,111726.46,106256.2,108444.03,102405.4
-2022-05-15,107705.21,111726.46,106256.2,108444.03,102405.4
-2022-05-16,109778.23,110471.7,107000.62,109279.98,102405.4
+2022-05-13,107705.21,111726.46,106256.2,108445.93,102405.4
+2022-05-14,107705.21,111726.46,106256.2,108445.93,102405.4
+2022-05-15,107705.21,111726.46,106256.2,108445.93,102405.4
+2022-05-16,109778.23,110471.7,107000.62,109281.88,102405.4
 2022-05-17,119372.81,125063.07,118465.59,123992.3,102405.4
 2022-05-18,124746.47,125463.37,121756.1,122484.67,102405.4
 2022-05-19,118696.67,120390.13,117178.96,117985.36,102405.4
@@ -1587,4 +1587,4 @@ date,open,high,low,close,holding
 2024-08-23,177144.9,179207.2,176091.8,177609.7,59914.74
 2024-08-24,177144.9,179207.2,176091.8,177609.7,59914.74
 2024-08-25,177144.9,179207.2,176091.8,177609.7,59914.74
-2024-08-26,178483.15,179979.85,177274.85,178829.1,59914.74
+2024-08-26,178483.15,179979.85,177274.85,178630.6,59914.74
