+++ DATA/GOLD/Holdings/Holdings_data.csv
@@ -3640,25 +3640,25 @@ date,segment,exchange,symbol,scrip_name,holding_quantity,avg_price,holding_amoun
 2022-05-12,EQ,NSE,TATAPOWER,TATA POWER,169.0,141.03,23834.0,221.11,37367.59,222.09,37533.21,212.31,35880.39,214.31,36218.39
 2022-05-12,EQ,NSE,VOLTAS,VOLTAS LTD,4.0,1059.0,4236.0,954.06,3816.24,957.66,3830.64,921.54,3686.16,935.73,3742.92
 2022-05-13,EQ,NSE,BPCL,BHARAT PETROLEUM CORPORATION L,43.0,457.0,19651.0,153.5,6600.5,155.25,6675.75,151.07,6496.01,151.5,6514.5
-2022-05-13,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1827.85,3655.7
+2022-05-13,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1828.8,3657.6
 2022-05-13,EQ,NSE,TATACHEM,TATA CHEMICALS LTD.,20.0,751.2,15024.0,904.25,18085.0,915.77,18315.4,873.71,17474.2,878.44,17568.8
 2022-05-13,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,383.92,38392.0,413.19,41319.0,383.92,38392.0,401.86,40186.0
 2022-05-13,EQ,NSE,TATAPOWER,TATA POWER,169.0,141.03,23834.0,219.69,37127.61,224.19,37888.11,216.47,36583.43,218.03,36847.07
 2022-05-13,EQ,NSE,VOLTAS,VOLTAS LTD,4.0,1059.0,4236.0,951.55,3806.2,951.55,3806.2,914.99,3659.96,917.99,3671.96
 2022-05-14,EQ,NSE,BPCL,BHARAT PETROLEUM CORPORATION L,43.0,457.0,19651.0,153.5,6600.5,155.25,6675.75,151.07,6496.01,151.5,6514.5
-2022-05-14,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1827.85,3655.7
+2022-05-14,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1828.8,3657.6
 2022-05-14,EQ,NSE,TATACHEM,TATA CHEMICALS LTD.,20.0,751.2,15024.0,904.25,18085.0,915.77,18315.4,873.71,17474.2,878.44,17568.8
 2022-05-14,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,383.92,38392.0,413.19,41319.0,383.92,38392.0,401.86,40186.0
 2022-05-14,EQ,NSE,TATAPOWER,TATA POWER,169.0,141.03,23834.0,219.69,37127.61,224.19,37888.11,216.47,36583.43,218.03,36847.07
 2022-05-14,EQ,NSE,VOLTAS,VOLTAS LTD,4.0,1059.0,4236.0,951.55,3806.2,951.55,3806.2,914.99,3659.96,917.99,3671.96
 2022-05-15,EQ,NSE,BPCL,BHARAT PETROLEUM CORPORATION L,43.0,457.0,19651.0,153.5,6600.5,155.25,6675.75,151.07,6496.01,151.5,6514.5
-2022-05-15,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1827.85,3655.7
+2022-05-15,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1828.8,3657.6
 2022-05-15,EQ,NSE,TATACHEM,TATA CHEMICALS LTD.,20.0,751.2,15024.0,904.25,18085.0,915.77,18315.4,873.71,17474.2,878.44,17568.8
 2022-05-15,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,383.92,38392.0,413.19,41319.0,383.92,38392.0,401.86,40186.0
 2022-05-15,EQ,NSE,TATAPOWER,TATA POWER,169.0,141.03,23834.0,219.69,37127.61,224.19,37888.11,216.47,36583.43,218.03,36847.07
 2022-05-15,EQ,NSE,VOLTAS,VOLTAS LTD,4.0,1059.0,4236.0,951.55,3806.2,951.55,3806.2,914.99,3659.96,917.99,3671.96
 2022-05-16,EQ,NSE,BPCL,BHARAT PETROLEUM CORPORATION L,43.0,457.0,19651.0,152.49,6557.07,152.71,6566.53,150.04,6451.72,151.11,6497.73
-2022-05-16,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1827.85,3655.7
+2022-05-16,EQ,NSE,LICI,LIFE INSURANCE CORPORATION,15.0,904.0,13560.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1828.8,3657.6
 2022-05-16,EQ,NSE,TATACHEM,TATA CHEMICALS LTD.,20.0,751.2,15024.0,884.22,17684.4,891.58,17831.6,872.61,17452.2,880.93,17618.6
 2022-05-16,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,408.47,40847.0,409.31,40931.0,393.76,39376.0,402.81,40281.0
 2022-05-16,EQ,NSE,TATAPOWER,TATA POWER,169.0,141.03,23834.0,220.82,37318.58,222.73,37641.37,215.58,36433.02,221.75,37475.75
@@ -6012,9 +6012,9 @@ date,segment,exchange,symbol,scrip_name,holding_quantity,avg_price,holding_amoun
 2024-08-25,EQ,NSE,TATAPOWER,TATA POWER,150.0,147.49,22124.0,423.0,63450.0,423.0,63450.0,417.4,62610.0,418.0,62700.0
 2024-08-25,MF,BMF,MIRAE_ASSET_ELSS_TAX_SAVER_FUND_DIRECT_PLAN_GROWTH,MIRAE ASSET TAX SAVER FUND - D,58.41,51.37,3000.37,660.89,0.0,664.27,0.0,659.61,0.0,660.89,0.0
 2024-08-25,MF,BMF,SBI_LONG_TERM_EQUITY_FUND_DIRECT_PLAN_GROWTH,SBI MAGNUM TAX GAIN SCHEME - D,6.64,452.01,3001.37,41.48,0.0,41.91,0.0,40.96,0.0,41.1,0.0
-2024-08-26,EQ,NSE,IRCTC,INDIAN RAILWAY CATERING AND TO,3.0,896.2,2688.6,924.75,2774.25,930.95,2792.85,924.75,2774.25,927.8,2783.4
-2024-08-26,EQ,NSE,KPITTECH,KPIT TECHNOLOGIES LTD,2.0,1500.0,3000.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1827.85,3655.7
-2024-08-26,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,1090.0,109000.0,1103.0,110300.0,1083.0,108300.0,1096.9,109690.0
-2024-08-26,EQ,NSE,TATAPOWER,TATA POWER,150.0,147.49,22124.0,420.1,63015.0,421.1,63165.0,417.0,62550.0,418.0,62700.0
+2024-08-26,EQ,NSE,IRCTC,INDIAN RAILWAY CATERING AND TO,3.0,896.2,2688.6,924.75,2774.25,930.95,2792.85,924.75,2774.25,928.5,2785.5
+2024-08-26,EQ,NSE,KPITTECH,KPIT TECHNOLOGIES LTD,2.0,1500.0,3000.0,1846.95,3693.9,1861.0,3722.0,1825.3,3650.6,1828.8,3657.6
+2024-08-26,EQ,NSE,TATAMOTORS,TATA MOTORS,100.0,261.0,26100.4,1090.0,109000.0,1103.0,110300.0,1083.0,108300.0,1093.6,109360.0
+2024-08-26,EQ,NSE,TATAPOWER,TATA POWER,150.0,147.49,22124.0,420.1,63015.0,421.1,63165.0,417.0,62550.0,418.85,62827.5
 2024-08-26,MF,BMF,MIRAE_ASSET_ELSS_TAX_SAVER_FUND_DIRECT_PLAN_GROWTH,MIRAE ASSET TAX SAVER FUND - D,58.41,51.37,3000.37,660.89,0.0,664.27,0.0,659.61,0.0,660.89,0.0
 2024-08-26,MF,BMF,SBI_LONG_TERM_EQUITY_FUND_DIRECT_PLAN_GROWTH,SBI MAGNUM TAX GAIN SCHEME - D,6.64,452.01,3001.37,41.48,0.0,41.91,0.0,40.96,0.0,41.1,0.0
