+++ DATA/SILVER/StockPrice/StockPrice_data.csv
@@ -772,7 +772,7 @@ date,symbol,open,high,low,close,volume
 2024-08-21,IRCTC,922.6,934.75,922.6,930.62,125995
 2024-08-22,IRCTC,933.96,938.24,930.02,935.15,35645
 2024-08-23,IRCTC,935.0,939.1,921.8,923.4,53542
-2024-08-26,IRCTC,924.75,930.95,924.75,927.8,79657
+2024-08-26,IRCTC,924.75,930.95,924.75,928.5,99231
 2024-06-03,KPITTECH,1514.96,1523.88,1436.12,1441.21,40559
 2024-06-04,KPITTECH,1456.07,1456.07,1292.96,1376.58,226624
 2024-06-05,KPITTECH,1404.06,1448.59,1322.43,1438.76,54197
@@ -829,7 +829,7 @@ date,symbol,open,high,low,close,volume
 2024-08-21,KPITTECH,1849.95,1855.0,1818.6,1830.15,50454
 2024-08-22,KPITTECH,1845.05,1868.1,1827.35,1830.7,20829
 2024-08-23,KPITTECH,1854.95,1854.95,1808.2,1824.75,20172
-2024-08-26,KPITTECH,1846.95,1861.0,1825.3,1827.85,10494
+2024-08-26,KPITTECH,1846.95,1861.0,1825.3,1828.8,12216
 2022-05-17,LICI,853.77,899.74,842.03,856.96,48792877
 2022-05-18,LICI,867.04,872.38,855.83,858.03,8919761
 2022-05-19,LICI,848.88,848.88,820.49,823.28,7372427
@@ -2766,7 +2766,7 @@ date,symbol,open,high,low,close,volume
 2024-08-21,TATAMOTORS,1091.4,1095.1,1080.0,1085.1,228030
 2024-08-22,TATAMOTORS,1090.7,1094.0,1065.05,1068.6,236226
 2024-08-23,TATAMOTORS,1071.8,1092.3,1071.0,1084.9,774166
-2024-08-26,TATAMOTORS,1090.0,1103.0,1083.0,1096.9,232571
+2024-08-26,TATAMOTORS,1090.0,1103.0,1083.0,1093.6,282346
 2021-02-01,TATAPOWER,74.31,78.42,71.99,77.16,2544260
 2021-02-02,TATAPOWER,79.63,81.27,77.02,80.79,2404196
 2021-02-03,TATAPOWER,81.22,82.57,79.43,81.37,4255011
@@ -3599,7 +3599,7 @@ date,symbol,open,high,low,close,volume
 2024-08-21,TATAPOWER,427.0,427.0,421.75,424.3,441380
 2024-08-22,TATAPOWER,426.0,426.05,421.9,422.85,199822
 2024-08-23,TATAPOWER,423.0,423.0,417.4,418.0,228889
-2024-08-26,TATAPOWER,420.1,421.1,417.0,418.0,420545
+2024-08-26,TATAPOWER,420.1,421.1,417.0,418.85,522246
 2021-06-01,VOLTAS,990.82,1000.13,986.7,993.61,619320
 2021-06-02,VOLTAS,991.31,1018.81,988.76,1014.4,1335354
 2021-06-03,VOLTAS,1021.85,1097.02,1016.9,1090.06,7137481
